shielddocs/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
boto3==1.34.0
openai==1.3.0
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.2
